+++ modules/imgproc/src/drawing.cpp
@@ -1184,7 +1184,7 @@ FillConvexPoly( Mat& img, const Point* v, int npts, const void* color, int line_
 
         edge[left].x = x1;
         edge[right].x = x2;
-        ptr += img.step;
+        ptr += (size_t)img.step;
     }
     while( ++y <= ymax );
 }
